# server/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PageMD EMR API",
        "version": "1.0.0",
        "description": "Commercial-grade REST API for PageMD Electronic Medical Records",
        "contact": {
            "name": "PageMD Support",
            "url": "https://pagemdemr.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://pagemdemr.com/terms"
        }
    },
    "servers": [
        {
            "url": "https://api.pagemdemr.com",
            "description": "Production"
        },
        {
            "url": "https://sandbox.pagemdemr.com",
            "description": "Sandbox"
        }
    ],
    "security": [
        {
            "oauth2": []
        }
    ],
...
    "paths": {
...
        "/api/v1/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "List patients",
                "operationId": "listPatients",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "updated_since",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of patients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Patient"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "patient.read"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Create a patient",
                "operationId": "createPatient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Patient"
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "patient.write"
                        ]
                    }
                ]
            }
        },
        "/api/v1/patients/{id}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Get patient by ID",
                "operationId": "getPatient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PatientDetail"
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "patient.read"
                        ]
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update patient",
                "operationId": "updatePatient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "patient.write"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "scopes": {
                            "patient.read": "Read patient records",
                            "patient.write": "Create/update patients",
                            "appointment.read": "Read appointments",
                            "appointment.write": "Create/update appointments",
                            "encounter.read": "Read encounters",
                            "encounter.write": "Create/update encounters",
                            "document.read": "Read documents",
                            "document.write": "Upload documents",
                            "webhook.manage": "Manage webhooks",
                            "ai.use": "Use AI features"
                        }
                    },
                    "clientCredentials": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        },
        "schemas": {
...
            "Patient": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "mrn": {
                        "type": "string"
                    },
                    "name": {
                        "type": "object",
                        "properties": {
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            },
                            "full": {
                                "type": "string"
                            }
                        }
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "sex": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PatientDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Patient"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "contact": {
                                "type": "object"
                            },
                            "address": {
                                "type": "object"
                            },
                            "emergency_contact": {
                                "type": "object"
                            },
                            "insurance": {
                                "type": "object"
                            }
                        }
                    }
                ]
            },
            "PatientCreate": {
                "type": "object",
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "sex": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "required": [
                    "first_name",
                    "last_name",
                    "birth_date"
                ]
            },
            "PatientUpdate": {
                "type": "object",
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "limit": {
                        "type": "integer"
                    },
                    "has_more": {
                        "type": "boolean"
                    },
                    "next_cursor": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "request_id": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "responses": {
...
            "Unauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - insufficient scope",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    }
}